minimal changes for dask run
