+++ aperturedb/ParallelLoader.py
@@ -21,17 +21,6 @@ class ParallelLoader(ParallelQuery.ParallelQuery):
             f"Starting ingestion with batchsize={batchsize}, numthreads={numthreads}")
         self.query(generator, batchsize, numthreads, stats)
 
-    def get_objects_existed(self):
-        return sum([stat["objects_existed"]
-                    for stat in self.actual_stats])
-
-    def get_suceeded_queries(self):
-        return sum([stat["suceeded_queries"]
-                    for stat in self.actual_stats])
-
-    def get_suceeded_commands(self):
-        return sum([stat["suceeded_commands"]
-                    for stat in self.actual_stats])
 
     def print_stats(self):
